Finally got the substitution of the newer grids correct
I added a new function for the default (16*16) grid and as soon as I moved the slider my oninput function calculated the fitting values and overwrote the existing grid in line 16, so that there is no problem with the stack overflowing when a new (customised) grid is created. I also added a button in my HTML file, but I actually had no need for it right now.